src/opencode/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "opencode-ai"
version = "1.0.0"
description = "Open source AI coding agent - Python edition. Guided by principles of balance, virtue, and ethical AI development."
readme = "README.md"
license = "MIT"
requires-python = ">=3.12"
authors = [
    { name = "OpenCode Community" }
]
keywords = ["ai", "coding", "agent", "llm", "assistant"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development",
    "Topic :: Text Editors :: Integrated Development Environments (IDE)",
    "Typing :: Typed",
]

dependencies = [
    # CLI
    "typer>=0.12.0",
    "rich>=13.0.0",
    
    # TUI
    "textual>=0.50.0",
    
    # Server
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "websockets>=12.0",
    
    # Database
    "sqlalchemy[asyncio]>=2.0.0",
    "aiosqlite>=0.20.0",
    "greenlet>=3.0.0",
    
    # HTTP Client
    "httpx>=0.27.0",
    
    # Validation
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    
    # Utilities
    "pyyaml>=6.0",
    "tomli>=2.0.0;python_version<'3.11'",
    "tomli-w>=1.0.0",
    "watchdog>=4.0.0",
    "pathspec>=0.12.0",
    "rapidfuzz>=3.0.0",
    "python-multipart>=0.0.9",
    "jinja2>=3.1.0",
    "packaging>=24.0",
    
    # Workflow Engine
    "apscheduler>=3.10.0",
    "jsonschema>=4.0.0",
    
    # GraphQL (optional - for workflow API)
    "strawberry-graphql>=0.220.0",
    
    # LLM Checker - Hardware detection
    "psutil>=5.9.0",
    
    # RAG Pipeline
    "numpy>=1.26.0",
    "scikit-learn>=1.3.0",
    
    # YouTube Transcript (optional)
    "youtube-transcript-api>=0.6.0",
    
    # Video Processing (optional - for multimodal RAG)
    # "moviepy>=1.0.3",
    # "SpeechRecognition>=3.10.0",
    
    # YouTube Data API (optional - for channel indexing)
    # "google-api-python-client>=2.116.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=5.0.0",
    "ruff>=0.5.0",
    "mypy>=1.10.0",
    "types-pyyaml>=6.0.0",
    "httpx-sse>=0.4.0",
]

# RAG Advanced Methods - Optional dependencies for advanced RAG features
rag-advanced = [
    "langchain>=0.3.0",
    "langchain-core>=0.3.0",
    "langchain-text-splitters>=0.3.0",
    "langchain-ollama>=0.3.0",
    "langchain-chroma>=0.2.0",
    "chromadb>=0.5.0",
    "ollama>=0.4.0",
]

# Fine-tuning - Optional dependencies for LLM fine-tuning
finetuning = [
    "torch>=2.0.0",
    "transformers>=4.40.0",
    "peft>=0.10.0",
    "accelerate>=0.30.0",
    "bitsandbytes>=0.43.0",
    "datasets>=2.18.0",
    "trl>=0.8.0",
]

# Multimodal - Optional dependencies for multimodal RAG
multimodal = [
    "pillow>=10.0.0",
    "transformers>=4.40.0",
]

# All optional dependencies
all = [
    "opencode-ai[dev,rag-advanced,finetuning,multimodal]",
]

[project.scripts]
opencode = "opencode.cli.main:app"

[project.urls]
Homepage = "https://opencode.ai"
Documentation = "https://opencode.ai/docs"
Repository = "https://github.com/opencode-ai/opencode-python"
Issues = "https://github.com/opencode-ai/opencode-python/issues"

[tool.hatch.build.targets.wheel]
packages = ["src/opencode"]

[tool.ruff]
target-version = "py312"
line-length = 100
src = ["src"]

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # Pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
    "ARG",    # flake8-unused-arguments
    "SIM",    # flake8-simplify
    "TCH",    # flake8-type-checking
    "PTH",    # flake8-use-pathlib
    "ERA",    # eradicate
    "RUF",    # Ruff-specific rules
]
ignore = [
    "E501",   # line too long (handled by formatter)
    "B008",   # do not perform function calls in argument defaults
    "B905",   # zip without explicit strict
    "ARG001", # unused function argument
    "ARG002", # unused method argument
    "PTH123", # use Path.open()
]

[tool.ruff.lint.isort]
known-first-party = ["opencode"]

[tool.mypy]
python_version = "3.12"
strict = true
warn_return_any = true
warn_unused_ignores = true
disallow_untyped_defs = true
plugins = ["pydantic.mypy"]

[[tool.mypy.overrides]]
module = [
    "watchdog.*",
    "pathspec.*",
    "rapidfuzz.*",
]
ignore_missing_imports = true

[tool.pytest.ini_options]
minversion = "8.0"
asyncio_mode = "auto"
testpaths = ["src/opencode/tests"]
pythonpath = ["src"]
python_classes = ["Test*"]
python_functions = ["test_*"]
python_files = ["test_*.py", "*_test.py"]
addopts = [
    "-ra",
    "-q",
    "--strict-markers",
    "--strict-config",
]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
    "ollama: marks tests that require Ollama server",
    "provider: marks tests for provider implementations",
    "prompt: marks tests for prompt evaluation",
    "e2e: marks end-to-end tests",
    "unit: marks unit tests",
]
filterwarnings = [
    "ignore::DeprecationWarning",
    "ignore::PendingDeprecationWarning",
]
log_cli = false
log_cli_level = "INFO"

[tool.coverage.run]
source = ["src/opencode"]
branch = true

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise NotImplementedError",
    "if TYPE_CHECKING:",
    "if __name__ == .__main__.:",
]
